consistently use int64, float64 and complex128
and use const x& instead of x const&

+++ include/cyten/config.h
@@ -14,19 +14,22 @@ class CytenConfig
     class PrintOptions
     {
       public:
-        int linewidth = 100;
-        int indent = 2;
-        int precision = 8; // #digits
-        int maxlines_spaces = 15;
-        int maxlines_tensors = 30;
-        bool skip_data = false;        // skip Data section in Tensor prints
-        bool summarize_blocks = false; // True -> always summarize (show only shape, not entries)
+        int64 linewidth = 100;
+        int64 indent = 2;
+        /// #digits
+        int64 precision = 8;
+        int64 maxlines_spaces = 15;
+        int64 maxlines_tensors = 30;
+        /// skip Data section in Tensor prints
+        bool skip_data = false;
+        /// True -> always summarize (show only shape, not entries)
+        bool summarize_blocks = false;
         PrintOptions() = default;
     };
 
     PrintOptions print_options = PrintOptions();
-    bool do_fusion_input_checks =
-      true; // If the symmetry methods should check their inputs are valid
+    /// If the symmetry methods should check their inputs are valid
+    bool do_fusion_input_checks = true;
     std::string default_symmetry_backend = "abelian";
     std::string default_block_backend = "numpy";
     CytenConfig() = default;
